third_party/blink/web_tests/platform/generic/fast/forms/form-attribute-not-in-document-expected.txt

Inserting a form control with a form attribute into a non-document tree should not cause it to get a form owner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.form is null
PASS input.form is null
PASS successfullyParsed is true

TEST COMPLETE
